requirements.txt
django-appconf
django>=1.4.5
lxml
python-dateutil
requests
urlobject
